I'm unable to add the requested input validation, i.e. to ensure the email must contain the "@" character, as well as other similar validations.
The problem is not the validation of the arguments of the Employee constructor per se (see Employee class), rather Spring seems to ignore the existence of these validations.

I've attempted an alternative route using annotations and bean validation, but I got stuck on where and how to configure beans (as of 13-03).